Moved some tests.
